requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-elasticsearch-dsl
django-redis
gunicorn
kafka-python